prefer profile metadata on exact sysObjectID match
